app/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
markdown2==2.4.10
gunicorn==21.2.0
logging-formatter-anticrlf==1.2
pytest==7.4.2
requests==2.31.0
textract==1.6.5
python-magic-bin==0.4.14
python-dotenv==1.1.0
